Changing to getResourceAsStream instead of new File()
This is to support accessing freemarker templates and config files from a jar instead of exploded file system.
I feel the implementation has to be opinionated towards packaging custom configurations in a jar that you can put on classpath/dependency rather than expecting a file deployment to any place the test have to run in.